refactor(ChatView): use form context for submission text
