+++ client/src/hooks/Messages/index.ts
@@ -1,5 +1,6 @@
 export { default as useAvatar } from './useAvatar';
 export { default as useProgress } from './useProgress';
+export { default as useSubmitMessage } from './useSubmitMessage';
 export { default as useMessageHelpers } from './useMessageHelpers';
 export { default as useCopyToClipboard } from './useCopyToClipboard';
 export { default as useMessageScrolling } from './useMessageScrolling';
